requirements.txt
openai
google-generativeai
python-dotenv
requests
PyGithub
pydantic
fastapi
uvicorn
python-multipart
